+++ Krowi_AchievementFilter.lua
@@ -131,8 +131,8 @@ function loadHelper:OnEvent(event, arg1, arg2)
          -- arg1 = isLogin, arg2 = isReload
         if arg1 or (arg2 and addon.Diagnostics.DebugEnabled()) then
             C_Timer.After(0, function()
-                C_Timer.After(addon.Options.db.EventReminders.PopUps.Show.OnLoginDelay, function()
-                    addon.GUI.AlertSystem.ShowActiveEvents();
+                C_Timer.After(addon.Options.db.EventReminders.OnLoginDelay, function()
+                    addon.GUI.AlertSystem.ShowActiveEventsOnLogin();
                 end);
             end);
         end