+++ Localization/deDE.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-01-13 15-03-17 ]] --
+-- [[ Exported at 2023-01-14 18-59-58 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --