+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Option to hide the completion date of the achievements
 - Option to delay the initial Event Reminders on login (default 5 seconds)
 - Movable Event Reminder Pop Ups
+- Event Reminders as Chat Messages
 
 ### Changed
 - Optimized the Event Handling
@@ -33,6 +34,7 @@ All notable changes to this project will be documented in this file.
 - Individual Dragon Racing achievements as they are now marked as tracking achievements
 - Start Time display for Event Reminders (This reduces the manual data collection on my end and has little to no importance)
 - Fated Raids Calendar Events
+- Going away with Dragonflight category
 
 ## 50.0 - 2022-12-26
 ### Added